Fix Firefox not catching images by xmlhttprequest.
In Firefox images that are loaded through XMLHttpRequest/Fetch are not
part of the "image" or "imageset" filter for WebRequest. Instead they
are labeld as "xmlhttprequest". To catch them we add the
"xmlhttprequest" to the filter and explicit check if the url is an
image. An example where this occurs is on a Discourse forum.